% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GcxGc_data_simplifier.R
\name{GcxGc_data_cleaning}
\alias{GcxGc_data_cleaning}
\title{Cleaning of the peaktables from the PEG4D Leco instrument}
\usage{
GcxGc_data_cleaning(x, sample_name, Blanks = TRUE, limit_of_detection = 3,
  simplify_matrix = TRUE, minimum.threshold = 50000, unknown = TRUE,
  column_bleed = TRUE, unlikely = TRUE, grouping = TRUE, RTspan = 40)
}
\arguments{
\item{x}{the peaktable obtained from the chromatof deconvolution software}

\item{sample_name}{the identificative letter/shortname of the samples. Mandatory to allow the algorithm to recognize the samples from the QCs.}

\item{Blanks}{should be the peaks delete according to their presence in blanks? If TRUE, every compound present also in the blanks with an average amount lower in the samples lower than the average amount found in the blanks multiplied by the LOD factor. default = TRUE}

\item{limit_of_detection}{the limit of detection for the compounds found also in the blanks. Default = 3}

\item{simplify_matrix}{if TRUE it will delete the compounds having a maximum value below the min.threshold value. default = TRUE}

\item{minimum.threshold}{the minimum threshold applied to reduce the matrix according to the maximum intensity found in any of the samples}

\item{unknown}{should unknown peaks be excluded? default set to TRUE}

\item{column_bleed}{should common column bleedings be excluded? default set to TRUE}

\item{unlikely}{should compounds containing unlikely atoms (chlorine, fluorine etc.) be excluded? default set to TRUE}

\item{grouping}{should compounds having the same name and similar retention times be grouped? Default set to TRUE}

\item{RTspan}{the maximum distance in seconds between to peaks with the same name to be grouped.}
}
\value{
a processed peaktable having all the real pekas present in your data in csv file in your working directory
}
\description{
The function transforms the raw peaktable obtained from the chromatof deconvolution software to a clean pekatable with only the names of the Blanks, QCs, samples, Retention times and compound names.
}
\examples{
Gar <- read.csv(system.file("extdata", "Gar.csv", package = "GCxGC.Leco.analyzer"), row.names = 1, stringsAsFactors = FALSE)
GcxGCdata <- GcxGc_data_cleaning(Gar)
}
\author{
Luca Narduzzi "nardluca@gmail.com"
}
